Add files via upload
Inheritance has been added to graph implementation. Now GraphMON, GraphLON, GraphTOE inherits from Graph. Code reuse.
