+++ GraphLON.h
@@ -1,46 +1,12 @@
 #pragma once
-#include <iostream>
-#include <cstdlib>
-#include <ctime>
-#include <vector>
-#include <stack>
+#include "Graph.h"
 
 
-class GraphLON
+class GraphLON : public Graph
 {
 private:
-	int numberOfNodes;
-	bool **graphMatrix;
 	std::vector<std::vector<int>> graphList;
 	std::vector<int> tmpList;
-
-	void randDAG()
-	{
-		graphMatrix = new bool *[numberOfNodes];
-		for (int i = 0; i < numberOfNodes; i++)
-		{
-			graphMatrix[i] = new bool[numberOfNodes];
-		}
-		for (int i = 0; i < numberOfNodes; i++)
-		{
-			for (int j = 0; j < numberOfNodes; j++)
-			{
-				graphMatrix[i][j] = false;
-			}
-		}
-		int numberOfEdges = (this->numberOfNodes *(this->numberOfNodes - 1)) / 4;
-		int row, col;
-		while (numberOfEdges > 0)
-		{
-			row = rand() % (this->numberOfNodes - 1);
-			col = rand() % (this->numberOfNodes - 1);
-			if (col > row && row < this->numberOfNodes - 1 && graphMatrix[row][col] != true)
-			{
-				graphMatrix[row][col] = true;
-				numberOfEdges--;
-			}
-		}
-	}
 
 	void transformMatrixToList()
 	{
@@ -58,15 +24,6 @@ private:
 		}
 	}
 
-	void deleteMatrix()
-	{
-		for (int i = 0; i < numberOfNodes; i++)
-		{
-			delete[] graphMatrix[i];
-		}
-		delete[] graphMatrix;
-	}
-
 public:
 
 	GraphLON()
@@ -76,24 +33,13 @@ public:
 
 	GraphLON(int numberOfNodes)
 	{
-		this->numberOfNodes = numberOfNodes;
-		randDAG();
+		newGraph(numberOfNodes);
 		transformMatrixToList();
-		deleteMatrix();
 	}
 
 	~GraphLON()
 	{
 		graphList.clear();
-	}
-
-	void newGraph(int numberOfNodes)
-	{
-		graphList.clear();
-		this->numberOfNodes = numberOfNodes;
-		randDAG();
-		transformMatrixToList();
-		deleteMatrix();
 	}
 
 	void printLON()
